+++ codigo-fonte/SpringApplication/src/main/java/com/system/application/modules/identity/collaborator/repository/CollaboratorRepository.java
@@ -12,7 +12,7 @@ import java.util.UUID;
 
 @Repository
 public interface CollaboratorRepository extends JpaRepository<Collaborator, UUID> {
-    long countBySchoolId(UUID schoolId);
+    int countBySchoolId(UUID schoolId);
 
     @Query("""
         SELECT c FROM Collaborator c
