+++ codigo-fonte/SpringApplication/src/main/java/com/system/application/modules/identity/legalguardian/repository/LegalGuardianRepository.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 @Repository
 public interface LegalGuardianRepository extends JpaRepository<LegalGuardian, UUID> {
     boolean existsByIdAndSchoolId(UUID legalGuardianId, UUID schoolId);
-    long countBySchoolId(UUID schoolId);
+    int countBySchoolId(UUID schoolId);
 
     @Query("""
         SELECT l FROM LegalGuardian l
