+++ codigo-fonte/SpringApplication/src/main/java/com/system/application/modules/school/dto/SchoolCapacityResponseDTO.java
@@ -1,8 +1,8 @@
 package com.system.application.modules.school.dto;
 
 public record SchoolCapacityResponseDTO(
-    long students,
-    long collaborators,
-    long legalGuardians,
-    long schoolAdmins
+    int students,
+    int collaborators,
+    int legalGuardians,
+    int schoolAdmins
 ) { }
